Move IntegerDataTypeInspectionTests for QuickFixes into relevant files under RD.QuickFixes
